1. Change wallet DB queries to be based on WalletSnapshot
So rather than each DB query being an action that reads the DB at that
point in the sequence of actions, the DB queries are all pure functions
of a WalletSnapshot. Then the only DB query action is
getWalletSnapshot :: m WalletSnapshot

This is a step towards ensuring that all read-only wallet APIs see a
single consistent snapshot of the wallet state.

This does not change the acid-state format, no migration is needed. But
it does require following patches because the API is different.

2. Update the wallet for the wallet DB query API changes

Many functions get passed in a WalletSnapshot. Actions that do DB
updates have to re-read a new snapshot. This should ultimately be
changed so that the updates return the new information, so that it is
atomic.